@@ -1,4 +1,4 @@
-package io.cygnuxltb.console.persistence.repository;
+package io.cygnuxltb.console.persistence.dao;
 
 import io.cygnuxltb.console.persistence.entity.TbtOrderExt;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author yellow013
  */
-public interface OrderExtRepository extends JpaRepository<TbtOrderExt, Long> {
+public interface OrderExtDao extends JpaRepository<TbtOrderExt, Long> {
 
 }
