@@ -1,4 +1,4 @@
-package io.cygnuxltb.console.persistence.repository;
+package io.cygnuxltb.console.persistence.dao;
 
 import io.cygnuxltb.console.persistence.entity.TbsPortfolio;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +12,7 @@ import java.util.List;
  * @author yellow013
  */
 @Repository
-public interface PortfolioRepository extends JpaRepository<TbsPortfolio, Long> {
+public interface PortfolioDao extends JpaRepository<TbsPortfolio, Long> {
 
 
     /**
