@@ -1,4 +1,4 @@
-package io.cygnuxltb.console.persistence.repository;
+package io.cygnuxltb.console.persistence.dao;
 
 import io.cygnuxltb.console.persistence.entity.TbtStrategy;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
  * @author yellow013
  */
 @Repository
-public interface StrategyRepository extends JpaRepository<TbtStrategy, Long> {
+public interface StrategyDao extends JpaRepository<TbtStrategy, Long> {
 
     /**
      * @param strategyId int
