@@ -1,8 +1,8 @@
 package io.cygnuxltb.console.service;
 
-import io.cygnuxltb.console.persistence.repository.AccountRepository;
-import io.cygnuxltb.console.persistence.entity.trade.AccountEntity;
-import io.cygnuxltb.console.service.util.DtoUtil;
+import io.cygnuxltb.console.persistence.dao.AccountDao;
+import io.cygnuxltb.console.persistence.entity.TbtAccount;
+import io.cygnuxltb.console.service.util.DtoConverter;
 import io.cygnuxltb.protocol.http.outbound.AccountDTO;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
@@ -16,13 +16,13 @@ import static io.cygnuxltb.console.persistence.util.DaoExecutor.select;
 public final class AccountService {
 
     @Resource
-    private AccountRepository repo;
+    private AccountDao dao;
 
     public List<AccountDTO> getAccount(int accountId) {
-        return select(AccountEntity.class,
-                () -> repo.queryByAccountId(accountId))
+        return select(TbtAccount.class,
+                () -> dao.queryByAccountId(accountId))
                 .stream()
-                .map(DtoUtil::convertToDTO)
+                .map(DtoConverter::toDTO)
                 .collect(Collectors.toList());
     }
 
