@@ -1,7 +1,7 @@
 package io.cygnuxltb.console.service;
 
-import io.cygnuxltb.console.persistence.entity.sys.UserEntity;
-import io.cygnuxltb.console.persistence.repository.UserRepository;
+import io.cygnuxltb.console.persistence.dao.UserDao;
+import io.cygnuxltb.console.persistence.entity.TbsUser;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 
@@ -11,10 +11,10 @@ import java.util.List;
 public final class UserService {
 
     @Resource
-    private UserRepository repo;
+    private UserDao dao;
 
     public boolean signIn(String sign, String password) {
-        List<UserEntity> list = repo.queryBy(sign, sign, sign, password);
+        List<TbsUser> list = dao.queryBy(sign, sign, sign, password);
         return list != null && list.size() > 0;
     }
 
